requirements.txt
requests>=2.25.0
python-telegram-bot>=20.0
APScheduler>=3.9.0
python-dotenv>=0.19.0
twilio>=8.0.0
openpyxl>=3.0.0